semiautonomous-agents/report-generator/requirements.txt
google-adk>=1.31.0
google-genai>=1.73.0
google-cloud-aiplatform[agent_engines]>=1.118.0
pydantic>=2.8
weasyprint>=62.0
jinja2>=3.1
markdown-it-py>=3.0
python-dotenv>=1.0
python-docx>=1.1
fastapi>=0.115
uvicorn[standard]>=0.30
sse-starlette>=2.1
httpx>=0.27
